Allow custom commit hash length
Fixes https://github.com/ember-cli-deploy/ember-cli-deploy-revision-data/issues/22

This allows the length of the commit hash that is used to be configured.
This applies to the `git-commit`, `git-tag-commit`, and `version-commit`
data generator types.

Example usage:

```js
ENV['revision-data'] = {
  type: 'git-commit',
  commitHashLength: 40
}
```

This example would cause the entire commit hash to be used as the
`revisionKey`.