[DOC] Better anchors in README
Now the README in the `type` section it says:

> *Default:* `'file-hash'`
> *Alternatives:* `'git-tag-commit'`, `'git-commit'`, `'version-commit'`

So I tend to cmd+f for `git-tag-commit`, but that string is nowhere else to be found in the README. So I have to do some mental parsing to find the correct section.

This will help people navigate around the README more quickly.